refactor!: decompose parser
