+++ package.json
@@ -31,7 +31,7 @@
     "qs": "^6.11.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.71.7",
+    "react-native": "0.71.8",
     "react-native-autoheight-webview": "^1.6.5",
     "react-native-get-random-values": "~1.8.0",
     "react-native-safe-area-context": "4.5.0",
