+++ parser/index.js
@@ -1,0 +1,19 @@
+import parseAbsenses from './absenses';
+import parseAnnounce from './announce';
+import parseMenu from './menu';
+import parseCurrentPoints from './signs';
+import parseShortTeachPlan from './teachPlan';
+import parseTeacherMessages from './teacherMessages';
+import parseTeachers from './teachers';
+import parseTimeTable from './timetable';
+
+export {
+  parseAbsenses,
+  parseAnnounce,
+  parseMenu,
+  parseCurrentPoints,
+  parseTeacherMessages,
+  parseTeachers,
+  parseShortTeachPlan,
+  parseTimeTable,
+};
