+++ utils/index.js
@@ -1,9 +1,7 @@
 import { downloadFile, saveFile, selectFile } from './files';
 import httpClient from './http';
-import DataParsing from './parser';
 import Storage from './storage';
 
 const storage = new Storage();
-const parser = new DataParsing();
 
-export { httpClient, storage, parser, downloadFile, saveFile, selectFile };
+export { httpClient, storage, downloadFile, saveFile, selectFile };
